+++ src/api/registration/CommandHandlerRegistration.ts
@@ -14,9 +14,7 @@
  * limitations under the License.
  */
 
-import { OnCommand } from "@atomist/automation-client/onCommand";
 import { NoParameters } from "@atomist/automation-client/SmartParameters";
-import { MachineOrMachineOptions } from "../../api-helper/machine/toMachineOptions";
 import { CommandListener } from "../listener/CommandListener";
 import { CommandRegistration } from "./CommandRegistration";
 
@@ -27,14 +25,8 @@ import { CommandRegistration } from "./CommandRegistration";
 export interface CommandHandlerRegistration<PARAMS = NoParameters> extends CommandRegistration<PARAMS> {
 
     /**
-     * Create the command function
-     * @return {AnyProjectEditor}
-     */
-    createCommand?: (sdm: MachineOrMachineOptions) => OnCommand<PARAMS>;
-
-    /**
      * Callback executing the command
      */
-    listener?: CommandListener<PARAMS>;
+    listener: CommandListener<PARAMS>;
 
 }